requirements.txt
fastapi==0.95.1
google-generativeai==0.1.0
joblib==1.3.2
numpy==1.24.3
scikit-learn==1.2.2
starlette==0.26.1
uvicorn==0.22.0
